add optional prefix/suffix for cron task names
Previously, the exact TaskDefinition name as specified in the crontab
would always be run. This made it difficult to use the same crontab
across live/test environments, as it is useful to use a separate set of
TaskDefinitions for different environments (for example, a "only note
what would be run" version in a test environment), while continuing to
test end-to-end connectivity.

--prefix=<prefix> adds a string to the beginning of the TaskDefinition
name, while --suffix=<suffix> adds a string to the end of the
TaskDefinition name