refactor Dockerfiles
previously, we had a separate "developement" and "release" Dockerfile.
The purpose of this was to reduce the size of the release image, at the
expense of giving it an external dependency (on the pre-compiled file
stored as part of the github release)

By adding the installation (and later, removal) of golang to the same
line as the build, we end up with an only-slightly-larger container
size, and a lot less complexity in our release process.

For one thing, this means that we don't need to tag/release each
individual feature in order to distribute it for testing.