Add a call-out when retrying tasks
previously, there was no indication that a task was being run due to a
previous failure - the output was identical regardless of how a task was
scheduled.

Add a new "Info" field to the TaskStatus struct, to allow the scheduler
to add additional, scheduler-specific information. In this case,
information about the number of attempts so-far and total.

Once the task is run, we inspect this information to provide additional
details about what is being run.