do not "retry" tasks which are still running
the intent was always that "failed" tasks would retry, while "still
running" tasks would be skipped, the logic being that these do not need
to be started - they are already started.

Due to bad logic in the code, however, these tasks were marked as
"failing" the same way any other error would be, resulting in a retry.
Most importantly, this meant that tasks which were running beyond their
scheduled time would often immediately restart, as soon as they
finished, resulting in a constant churn for tasks which were meant to
warn of their long runs, and skip the next run to compensate. This was
all intended in the name of cluster resource conservation, but the error
had exactly the opposite intended effect.

To correct this, we add a new "Running" part to the TaskStatus struct.
This indicates whether or not we know the task to be "already running".
False does not necessarily indicate that the task is not running, but
True indicates that the task is known to have run recently-enough that
we do not expect it should be run immediately (ie: it was still running
when we looked).

The ECSTaskRunner is now split into a "TaskRunner" and a separate
"SkipRunning" TaskRunner. The SkipRunning taskrunner examines ECS prior
to passing results back into the "real" taskrunner. No sanity-check is
made to ensure the clustername is the same between the two - this is
because one could theoretically chain multiple SkipRunning taskrunners
to check multiple clusters, prior to running a task on a specific
cluster. The SkipRunning TaskRunner is the one that potentially sets the
"Running" indicator, which the Retry scheduler now tests before marking
a task as needing a retry.