add a -dump option, to output the schedule
a request was made to list a summary of the scheduled tasks, so that
the schedule could be compared with (external) monitoring data, in order
to detect any discrepancies.

To achieve this, we add a generic Dump() method, for outputting a
summary of scheduled events to a channel, and a DumpJson() method, for
formatting that summary information as a JSON array.

The command-line gains a -dump parameter, as well as additional
parameters for controlling the beginning/end of the range to dump. A
-dump-format=<format> parameter is also added, but only accepts "json"
at this time.

The name "Dump()" for the channel-returning method is very specific to
this request, but the functionality is potentially generically-useful,
and may even be a suitable interface for some scheduling workflows (for
example, if Tick() were not being used directly). This would not be
suitable for internal use, as it is not in keeping with ecscron's goals
of having a stable memory footprint. That said, it may be worth thinking
about, and may be renamed at some point if it is deemed useful.

The new Dump() code does not have 100% test coverage. Specifically,
there is nothing testing the error-detection of the JSON Marshaller or
failed io Writes.