allow the use of SIGUSR1 to pause/resume
During certain operations, it is useful to temporarily suspend cron. For
example, during deploys or scaling events. Originally, -async was
intended to fit this situation, ie: the cron process would be stopped,
and would later be re-started with -async to account for the gap in time
when cron was disabled.

This had a couple of problems:

 - starting and stopping the cron process can be problematic to manage.
   In particular, bringing the process down completely may easily trigger
   alarms or auto-healing, meaning one generally needs to coordinate
   with some external system, rather than simply "killing and later
   re-starting"

 - Killing the process may easily stop things while in the middle of
   processing a tick, while a "pause" should really only happen between
   ticks, to ensure consistency.

Add a listener for SIGUSR1, allowing cron to be paused/resumed, without
going as far as to bring down the whole cron process. This signal is
only processed between ticks, so there shouldn't be any "half run"
processes, or processes which are run but not reported, etc.