teach -pause option, for starting in "paused" mode
While "pausing" is a nice alternative to stopping and restarting the
cron process, it is not applicable to all situations. For example, one
may wish to pause the cron process, and later replace it prior to being
ready to resume it. For example, if both cron and another process are
being deployed at the same time, you may wish to pause cron for a
deploy, and then the deploy also involves restarting the cron process.

In such situations, you may want cron to "start paused", awaiting a
SIGUSR1 before it will resume in earnest.

We add a -pause option for this situation.