teach -max-pause option, a safety hatch for pauses
Previously, any "pause", or initialization with "-pause", would wait
eternally for a SIGUSR1 to resume. This is a good default, by the
principle of "least surprise", but it is not always desireable.

If, for whatever reason, a signal is not sent (or received), there was
no fallback to ensure that cron *eventually* resumes.

Teach the -max-pause=<duration> option, for ensuring that no pause lasts
longer then the specified duration.

The specific use-case this was intended to resolve was a deploy script
which paused cron during deploy, but could theoretically die prior to
completion. Rather than relying on an external process to monitor the
state of cron, telling cron "never pause for longer than 10 minutes",
for example, keeps things simple. Of course, with the caveat that longer
pauses will always require one to terminate the process.